// types.ts
export type TransactionType = {
     id?: string;
     type: string;
     amount: number;
     category?: string;
     date: Date | Timestamp | string;
     description?: string;
     image?: any;
     uid?: string;
     walletId: string;
};
// ...
export type WalletType = {
     id?: string;
     name: string;
     amount?: number;
     totalIncome?: number;
     totalExpenses?: number;
     image: any;
     uid?: string;
     created?: Date;
};